# PERF: Increase threashold for using binary search in IndexEngine