--- a/doc/source/whatsnew/v2.2.0.rst
+++ b/doc/source/whatsnew/v2.2.0.rst
@@ -105,7 +105,7 @@ Deprecations
 
 Performance improvements
 ~~~~~~~~~~~~~~~~~~~~~~~~
--
+- Increase the threshold to use binary search instead of a linear search in IndexEngine :issue:`54550`
 -
 
 .. ---------------------------------------------------------------------------