--- a/pandas/_libs/index.pyx
+++ b/pandas/_libs/index.pyx
@@ -375,7 +375,7 @@ cdef class IndexEngine:
         # map each starget to its position in the index
         if (
                 stargets and
-                len(stargets) < 5 and
+                len(stargets) < (n / (2 * n.bit_length())) and
                 not na_in_stargets and
                 self.is_monotonic_increasing
         ):